Use a method out of react scope to set CSS vars as that doesnt need react
