+++ apps/web/public/embed-init-iframe.js
@@ -11,7 +11,7 @@ window.resetEmbedStatus = () => {
   try {
     // eslint-disable-next-line @calcom/eslint/avoid-web-storage
     window.sessionStorage.removeItem("calEmbedMode");
-  } catch (e) {}
+  } catch (e) { }
 };
 
 window.getEmbedTheme = () => {
@@ -24,3 +24,23 @@ window.getEmbedNamespace = () => {
   const namespace = url.searchParams.get("embed");
   return namespace;
 };
+window.CalEmbed = window.CalEmbed || {};
+window.CalEmbed.applyCssVars = (cssVarsPerTheme) => {
+  const cssVarsStyle = [];
+  if (cssVarsPerTheme) {
+    for (const [themeName, cssVars] of Object.entries(cssVarsPerTheme)) {
+      cssVarsStyle.push(`.${themeName} {`);
+      for (const [cssVarName, value] of Object.entries(cssVars)) {
+        cssVarsStyle.push(`--${cssVarName}: ${value};`);
+      }
+      cssVarsStyle.push(`}`);
+    }
+  }
+  const style = document.createElement("style");
+  style.id = "embed-css-vars"
+  style.innerText = cssVarsStyle.join("\n");
+  if (document.head.querySelector("#embed-css-vars")) {
+    return;
+  }
+  document.head.appendChild(style);
+}
